+++ Directorio/Client/src/modelo_d/echo/Conexion.java
@@ -1,4 +1,4 @@
-package echo;
+package modelo_d.echo;
 
 import java.io.IOException;
 import java.io.OutputStream;
