+++ Directorio/Client/src/modelo_d/echo/Conexiones.java
@@ -1,4 +1,4 @@
-package echo;
+package modelo_d.echo;
 
 
 import java.io.IOException;
